backend/requirements.txt
# Backend API
fastapi==0.109.0
uvicorn==0.27.0

# Dashboard Dash (Monitoramento.py)
dash==2.14.2
plotly==5.18.0

# Dados
pandas==2.2.0
gspread==6.0.0
google-auth==2.27.0
